backend/requirements.txt
asgiref==3.9.1
channels==4.3.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
channels
channels_redis # optional (for production) — local memory layer ok for dev
psycopg2-binary
python-dotenv